Refactor: Move isolate UUID from jsg::Lock to Worker::Isolate.
It turns out the only production use of this API is at `Worker::Actor` creation time, and only for metrics purposes. I would like to eliminate the need for holding the isolate lock when constructing an Actor, but then I can't get the UUID. Moving the UUID to be stored on the Worker::Isolate object makes it easy to get at without a lock.